Some tidying up, refactoring and additional comments
