+++ WpfTimelineControl/TimelineEntryColorConverters.cs
@@ -42,10 +42,10 @@ namespace WpfTimelineControl
             var brushColor = foreground.Color;
 
             var colorInDrawing = System.Drawing.Color.FromArgb(brushColor.A, brushColor.R, brushColor.G, brushColor.B);
-            ColorHelper.ColorToHsv(colorInDrawing, out double h, out _, out _);
+            ColorConverter.ColorToHsv(colorInDrawing, out double h, out _, out _);
 
             // Mimicking the same hue at 0.05% opacity over white, but opaque
-            var backgroundInDrawing = ColorHelper.HsvToColor(h, 0.08, 0.98);
+            var backgroundInDrawing = ColorConverter.HsvToColor(h, 0.08, 0.98);
             return new SolidColorBrush(Color.FromArgb(backgroundInDrawing.A, backgroundInDrawing.R, backgroundInDrawing.G, backgroundInDrawing.B));
         }
