+++ WpfTimelineControl/ViewModelBase.cs
@@ -1,0 +1,46 @@
+﻿using System;
+using System.ComponentModel;
+using System.Runtime.CompilerServices;
+using System.Windows.Input;
+
+namespace WpfTimelineControl
+{
+    public abstract class ViewModelBase : INotifyPropertyChanged
+    {
+        public event PropertyChangedEventHandler PropertyChanged;
+
+        protected void NotifyPropertyChanged([CallerMemberName] string propertyName = "")
+        {
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+        }
+    }
+
+    public class MyCommand : ICommand
+    {
+        private readonly Action action;
+        private readonly Func<bool> canExecute;
+
+        public MyCommand(Action action, Func<bool> canExecute = null)
+        {
+            this.action = action;
+            this.canExecute = canExecute;
+        }
+
+        public event EventHandler CanExecuteChanged;
+
+        public bool CanExecute(object parameter)
+        {
+            return canExecute?.Invoke() ?? true;
+        }
+
+        public void Execute(object parameter)
+        {
+            action?.Invoke();
+        }
+
+        public void RaiseCanExecuteChanged()
+        {
+            CanExecuteChanged?.Invoke(this, new EventArgs());
+        }
+    }
+}
