+++ WpfTimelineControl/WpfTimelineControl.csproj
@@ -49,8 +49,9 @@
     <Reference Include="PresentationFramework" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="TimelineModels.cs" />
     <Compile Include="BooleanConverter.cs" />
-    <Compile Include="ColorHelper.cs" />
+    <Compile Include="ColorConverter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs">
       <SubType>Code</SubType>
     </Compile>
@@ -70,10 +71,10 @@
     <Compile Include="TimelineBuilder.cs" />
     <Compile Include="TimelineConstants.cs" />
     <Compile Include="TimelineElementMarginConverter.cs" />
-    <Compile Include="TimelineElements.cs" />
     <Compile Include="TimelineElementWidthConverter.cs" />
     <Compile Include="TimelineEntryColorConverters.cs" />
     <Compile Include="TimelineViewModel.cs" />
+    <Compile Include="ViewModelBase.cs" />
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
